챗지피티2

+++ src/main/java/reduck/reduck/domain/chatgpt/dto/GptUsableCountResponse.java
@@ -6,5 +6,5 @@ import lombok.Getter;
 @Builder
 @Getter
 public class GptUsableCountResponse {
-    private int remainUsageCount;
+    private Long remainUsageCount;
 }
